[feat] 재발급 에러처리 적용 확인 테스트
