--- src/main/java/modernfarmer/server/farmususer/global/exception/unauthorized/InvalidBearerException.java
@@ -1,8 +1,0 @@
-package modernfarmer.server.farmususer.global.exception.unauthorized;
-
-public class InvalidBearerException extends UnauthorizedException {
-
-    public InvalidBearerException() {
-        super("로그인이 필요한 서비스입니다.", 1013);
-    }
-}
