+++ src/main/java/modernfarmer/server/farmususer/user/service/AuthService.java
@@ -145,72 +145,6 @@ public class AuthService{
     }
 
 
-    public ResponseDto logout(Long userId) {
-        deleteValueByKey(String.valueOf(userId));
-
-        ResponseDto responseDto = ResponseDto.builder()
-                .message("OK")
-                .code(200)
-                .build();
-        return responseDto;
-    }
-
-    public TokenResponseDto reissueToken(String refreshToken, Long userId) {
-        TokenResponseDto reissueTokenResponse;
-
-        if(!jwtTokenProvider.validateRefreshToken(refreshToken)){
-
-            reissueTokenResponse = TokenResponseDto.builder()
-                    .code(417)
-                    .message("재로그인하시오")
-                    .build();
-            
-
-            new TokenExpiredException();
-
-            return reissueTokenResponse;
-        }
-
-        String redisRefreshToken = redisTemplate.opsForValue().get(userId.toString());
-
-        if(redisRefreshToken.equals(refreshToken)){
-
-            String userRole = userRepository.findUserRole(userId);
-
-            reissueTokenResponse= TokenResponseDto
-                    .builder()
-                    .code(200)
-                    .message("OK")
-                    .accessToken(jwtTokenProvider.createAccessToken(userId, userRole))
-                    .refreshToken(refreshToken)
-                    .build();
-
-            return reissueTokenResponse;
-
-        }
-
-        reissueTokenResponse = TokenResponseDto.builder()
-                .code(403)
-                .message("접근이 올바르지 않습니다.")
-                .build();
-
-        return reissueTokenResponse;
-
-    }
-
-
-
-
-    public void deleteValueByKey(String key) {
-        redisTemplate.delete(key);
-    }
-
-
-
-
-
-
-
     public Mono<KakaoUserResponseDto> getUserKakaoInfo(String accessToken) {
         return webClient
                 .get()
