+++ src/main/java/modernfarmer/server/farmususer/user/util/JwtTokenProvider.java
@@ -26,10 +26,10 @@ public class JwtTokenProvider {
     private String secretKey;
 
 
-    private final long accessTokenTime = 30L * 24L * 60 * 60 * 1000; // 1달 토큰 유효
+    private final long accessTokenTime = 30L * 1000;; // 1달 토큰 유효
 
 
-    private final long refreshTokenTime = 1L * 60 * 1000 * 2; // 1달 토큰 유효
+    private final long refreshTokenTime = 30L * 1000 * 2; // 1달 토큰 유효
 
     @PostConstruct
     protected void init() {
